Move shadertoy support directly into renderer via control messages.
