+++ iris/renderer.cc
@@ -3,17 +3,6 @@
  */
 #include "config.h"
 
-#include "absl/container/flat_hash_map.h"
-#if PLATFORM_COMPILER_GCC
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wmaybe-uninitialized"
-#endif
-#include "absl/container/inlined_vector.h"
-#if PLATFORM_COMPILER_GCC
-#pragma GCC diagnostic pop
-#endif
-#define _TURN_OFF_PLATFORM_STRING // disable U macro in cpprest
-#include "cpprest/http_client.h"
 #include "enumerate.h"
 #include "error.h"
 #include "glm/common.hpp"
@@ -23,9 +12,9 @@
 #include "glslang/Public/ShaderLang.h"
 #include "gsl/gsl"
 #include "io/json.h"
+#include "io/shadertoy.h"
 #include "protos.h"
-#include "renderer.h"
-#include "string_util.h"
+#include "renderer_util.h"
 #if PLATFORM_COMPILER_MSVC
 #pragma warning(push)
 #pragma warning(disable : 4127)
@@ -49,13 +38,10 @@
 #include <chrono>
 #include <cstdlib>
 #include <exception>
-#include <fstream>
 #include <functional>
 #include <memory>
-#include <mutex>
 #include <string>
 #include <utility>
-#include <vector>
 
 using namespace std::string_literals;
 
@@ -100,41 +86,39 @@ GetLogger(spdlog::sinks_init_list logSinks = {}) noexcept {
 
 namespace iris::Renderer {
 
-static bool sRunning{false};
+VkInstance sInstance{VK_NULL_HANDLE};
+VkDebugUtilsMessengerEXT sDebugUtilsMessenger{VK_NULL_HANDLE};
+VkPhysicalDevice sPhysicalDevice{VK_NULL_HANDLE};
+VkDevice sDevice{VK_NULL_HANDLE};
+VmaAllocator sAllocator{VK_NULL_HANDLE};
 
-static VkInstance sInstance{VK_NULL_HANDLE};
-static VkDebugUtilsMessengerEXT sDebugUtilsMessenger{VK_NULL_HANDLE};
-static VkPhysicalDevice sPhysicalDevice{VK_NULL_HANDLE};
-static VkDevice sDevice{VK_NULL_HANDLE};
-static VmaAllocator sAllocator{VK_NULL_HANDLE};
+std::uint32_t sGraphicsQueueFamilyIndex{UINT32_MAX};
+absl::InlinedVector<VkQueue, 16> sGraphicsCommandQueues;
+absl::InlinedVector<VkCommandPool, 16> sGraphicsCommandPools;
+absl::InlinedVector<VkFence, 16> sGraphicsCommandFences;
 
-static std::uint32_t sGraphicsQueueFamilyIndex{UINT32_MAX};
-static absl::InlinedVector<VkQueue, 16> sGraphicsCommandQueues;
-static absl::InlinedVector<VkCommandPool, 16> sGraphicsCommandPools;
-static absl::InlinedVector<VkFence, 16> sGraphicsCommandFences;
+VkRenderPass sRenderPass{VK_NULL_HANDLE};
 
-static VkRenderPass sRenderPass{VK_NULL_HANDLE};
+std::uint32_t const sNumRenderPassAttachments{4};
+std::uint32_t const sColorTargetAttachmentIndex{0};
+std::uint32_t const sColorResolveAttachmentIndex{1};
+std::uint32_t const sDepthStencilTargetAttachmentIndex{2};
+std::uint32_t const sDepthStencilResolveAttachmentIndex{3};
 
-static constexpr std::uint32_t const sNumRenderPassAttachments{4};
-static constexpr std::uint32_t const sColorTargetAttachmentIndex{0};
-static constexpr std::uint32_t const sColorResolveAttachmentIndex{1};
-static constexpr std::uint32_t const sDepthStencilTargetAttachmentIndex{2};
-static constexpr std::uint32_t const sDepthStencilResolveAttachmentIndex{3};
+VkSurfaceFormatKHR const sSurfaceColorFormat{VK_FORMAT_B8G8R8A8_UNORM,
+                                             VK_COLOR_SPACE_SRGB_NONLINEAR_KHR};
+VkFormat const sSurfaceDepthStencilFormat{VK_FORMAT_D32_SFLOAT};
+VkSampleCountFlagBits const sSurfaceSampleCount{VK_SAMPLE_COUNT_4_BIT};
+VkPresentModeKHR const sSurfacePresentMode{VK_PRESENT_MODE_FIFO_KHR};
 
-static constexpr VkSurfaceFormatKHR const sSurfaceColorFormat{
-  VK_FORMAT_B8G8R8A8_UNORM, VK_COLOR_SPACE_SRGB_NONLINEAR_KHR};
-static constexpr VkFormat const sSurfaceDepthStencilFormat{
-  VK_FORMAT_D32_SFLOAT};
-static constexpr VkSampleCountFlagBits const sSurfaceSampleCount{
-  VK_SAMPLE_COUNT_4_BIT};
-static constexpr VkPresentModeKHR const sSurfacePresentMode{
-  VK_PRESENT_MODE_FIFO_KHR};
-
-static absl::flat_hash_map<std::string, iris::Window>& Windows() {
+absl::flat_hash_map<std::string, iris::Window>& Windows() {
   static absl::flat_hash_map<std::string, iris::Window> sWindows;
   return sWindows;
 } // Windows
 
+Renderables sRenderables;
+
+static bool sRunning{false};
 static bool sInFrame{false};
 static std::uint32_t sFrameNum{0};
 static std::chrono::steady_clock::time_point sPreviousFrameTime{};
@@ -149,92 +133,6 @@ static tbb::task_scheduler_init sTaskSchedulerInit{
   tbb::task_scheduler_init::deferred};
 static tbb::concurrent_queue<std::function<std::system_error(void)>>
   sIOContinuations{};
-
-class Renderables {
-public:
-  std::vector<Component::Renderable> operator()() {
-    std::lock_guard<std::mutex> lock{mutex_};
-    return renderables_;
-  }
-
-  void push_back(Component::Renderable renderable) {
-    std::lock_guard<std::mutex> lock{mutex_};
-    renderables_.clear();
-    renderables_.push_back(std::move(renderable));
-  }
-
-private:
-  std::mutex mutex_{};
-  std::vector<Component::Renderable> renderables_{};
-}; // class Renderables
-
-static Renderables sRenderables;
-
-static VkCommandBuffer
-Render(iris::Renderer::Component::Renderable const& renderable,
-       VkViewport* pViewport, VkRect2D* pScissor) noexcept {
-  VkCommandBufferAllocateInfo commandBufferAI = {};
-  commandBufferAI.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-  commandBufferAI.commandPool = sGraphicsCommandPools[0];
-  commandBufferAI.level = VK_COMMAND_BUFFER_LEVEL_SECONDARY;
-  commandBufferAI.commandBufferCount = 1;
-
-  VkCommandBuffer commandBuffer;
-  if (auto result =
-        vkAllocateCommandBuffers(sDevice, &commandBufferAI, &commandBuffer);
-      result != VK_SUCCESS) {
-    GetLogger()->error("Cannot allocate command buffer: {}", to_string(result));
-    return VK_NULL_HANDLE;
-  }
-
-  VkCommandBufferInheritanceInfo commandBufferII = {};
-  commandBufferII.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_INHERITANCE_INFO;
-  commandBufferII.renderPass = sRenderPass;
-  commandBufferII.subpass = 0;
-  commandBufferII.framebuffer = VK_NULL_HANDLE;
-
-  VkCommandBufferBeginInfo commandBufferBI = {};
-  commandBufferBI.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-  commandBufferBI.flags = VK_COMMAND_BUFFER_USAGE_SIMULTANEOUS_USE_BIT |
-                          VK_COMMAND_BUFFER_USAGE_RENDER_PASS_CONTINUE_BIT;
-  commandBufferBI.pInheritanceInfo = &commandBufferII;
-
-  vkBeginCommandBuffer(commandBuffer, &commandBufferBI);
-  vkCmdBindPipeline(commandBuffer, VK_PIPELINE_BIND_POINT_GRAPHICS,
-                    renderable.pipeline);
-
-  vkCmdSetViewport(commandBuffer, 0, 1, pViewport);
-  vkCmdSetScissor(commandBuffer, 0, 1, pScissor);
-
-  if (renderable.descriptorSet != VK_NULL_HANDLE) {
-    vkCmdBindDescriptorSets(commandBuffer, VK_PIPELINE_BIND_POINT_GRAPHICS,
-                            renderable.pipelineLayout, 0, 1,
-                            &renderable.descriptorSet, 0, nullptr);
-  }
-
-  if (renderable.vertexBuffer != VK_NULL_HANDLE) {
-    vkCmdBindVertexBuffers(commandBuffer, 0, 1, &renderable.vertexBuffer,
-                           &renderable.vertexBufferBindingOffset);
-  }
-
-  if (renderable.indexBuffer != VK_NULL_HANDLE) {
-    vkCmdBindIndexBuffer(commandBuffer, renderable.indexBuffer,
-                         renderable.indexBufferBindingOffset,
-                         renderable.indexType);
-  }
-
-  if (renderable.numIndices > 0) {
-    vkCmdDrawIndexed(commandBuffer, renderable.numIndices,
-                     renderable.instanceCount, renderable.firstIndex,
-                     renderable.vertexOffset, renderable.firstInstance);
-  } else {
-    vkCmdDraw(commandBuffer, renderable.numVertices, renderable.instanceCount,
-              renderable.firstVertex, renderable.firstInstance);
-  }
-
-  vkEndCommandBuffer(commandBuffer);
-  return commandBuffer;
-} // Render
 
 static void
 CreateEmplaceWindow(iris::Control::Window const& windowMessage) noexcept {
@@ -261,324 +159,70 @@ CreateEmplaceWindow(iris::Control::Window const& windowMessage) noexcept {
   }
 } // CreateEmplaceWindow
 
-static void LoadShaderToy(iris::Control::ShaderToy const& shaderToy) noexcept {
-  IRIS_LOG_ENTER();
+static VKAPI_ATTR VkBool32 VKAPI_CALL DebugUtilsMessengerCallback(
+  VkDebugUtilsMessageSeverityFlagBitsEXT messageSeverity,
+  VkDebugUtilsMessageTypeFlagsEXT messageTypes,
+  VkDebugUtilsMessengerCallbackDataEXT const* pCallbackData, void*) {
+  using namespace std::string_literals;
 
-  class LoadTask : public tbb::task {
-  public:
-    LoadTask(iris::Control::ShaderToy const& st) {
-      for (int i = 0; i < st.url_size(); ++i) { urls_.push_back(st.url(i)); }
+  fmt::memory_buffer buf;
+  fmt::format_to(
+    buf, "{}: {}",
+    to_string(static_cast<VkDebugUtilsMessageTypeFlagBitsEXT>(messageTypes)),
+    pCallbackData->pMessage);
+  std::string const msg(buf.data(), buf.size());
+
+  buf.clear();
+  for (uint32_t i = 0; i < pCallbackData->objectCount; ++i) {
+    if (pCallbackData->pObjects[i].pObjectName) {
+      fmt::format_to(buf, "{}, ", pCallbackData->pObjects[i].pObjectName);
     }
+  }
+  std::string const objNames(buf.data(), buf.size() == 0 ? 0 : buf.size() - 2);
 
-    tbb::task* execute() override {
-      IRIS_LOG_ENTER();
-      if (urls_.empty()) {
-        GetLogger()->debug("No ShaderToy URLs in message");
-        IRIS_LOG_LEAVE();
-        return nullptr;
-      }
-
-      if (urls_.size() > 1) {
-        GetLogger()->error("Multiple ShaderToy URLs not implemented");
-        IRIS_LOG_LEAVE();
-        return nullptr;
-      }
-
-#if PLATFORM_WINDOWS
-      web::http::uri const viewURI(string_to_wstring(urls_[0]));
-#else
-      web::http::uri const viewURI(urls_[0]);
-#endif
-
-      // grab the last component of the uri path: that's the shaderID
-      auto const path = viewURI.path();
-      auto const id = path.find_last_of('/');
-
-#if PLATFORM_WINDOWS
-      if (id == std::wstring::npos) {
-#else
-      if (id == std::string::npos) {
-#endif
-        GetLogger()->error("Bad URL: {}", urls_[0]);
-        IRIS_LOG_LEAVE();
-        return nullptr;
-      }
-
-      web::http::uri_builder apiURI;
-      apiURI.set_scheme(viewURI.scheme());
-      apiURI.set_host(viewURI.host());
-      apiURI.set_path(_XPLATSTR("api/v1/shaders"));
-      apiURI.append_path(path.substr(id));
-      apiURI.append_query(_XPLATSTR("key=BtHKWW"));
-#if PLATFORM_WINDOWS
-      GetLogger()->debug("api URI: {}", wstring_to_string(apiURI.to_string()));
-#else
-      GetLogger()->debug("api URI: {}", apiURI.to_string());
-#endif
-
-      std::string const code = GetCode(apiURI.to_uri());
-  
-      Component::Renderable renderable;
-
-      auto vs = iris::Renderer::CompileShaderFromSource(
-        R"(#version 450
-layout(push_constant) uniform uPC {
-    vec4 iMouse;
-    float iTime;
-    float iTimeDelta;
-    float iFrameRate;
-    float iFrame;
-    vec3 iResolution;
-    float padding0;
-};
-
-layout(location = 0) out vec2 fragCoord;
-
-void main() {
-    fragCoord = vec2((gl_VertexIndex << 1) & 2, (gl_VertexIndex & 2));
-    gl_Position = vec4(fragCoord * 2.0 - 1.0, 0.f, 1.0);
-    // flip to match shadertoy
-    fragCoord.y *= -1;
-    fragCoord.y += 1;
-
-    // multiple by resolution to match shadertoy
-    fragCoord *= iResolution.xy;
-})",
-        VK_SHADER_STAGE_VERTEX_BIT, "Renderable::ShaderToyVertexShader");
-      if (!vs) {
-        GetLogger()->error("Cannot create ShaderToy vertex shader: {}",
-                           vs.error().what());
-        IRIS_LOG_LEAVE();
-        return nullptr;
-      }
-
-      std::ostringstream fragmentShaderSource;
-      fragmentShaderSource <<
-        R"(#version 450
-#extension GL_GOOGLE_include_directive : require
-layout(push_constant) uniform uPC {
-    vec4 iMouse;
-    float iTime;
-    float iTimeDelta;
-    float iFrameRate;
-    float iFrame;
-    vec3 iResolution;
-    float padding0;
-};
-
-layout(location = 0) in vec2 fragCoord;
-layout(location = 0) out vec4 fragColor;
-)" << code << R"(
-
-void main() {
-    mainImage(fragColor, fragCoord);
-})";
-
-  auto fs = iris::Renderer::CompileShaderFromSource(
-        fragmentShaderSource.str(), VK_SHADER_STAGE_FRAGMENT_BIT,
-        "iris-shadertoy::Renderable::FragmentShader");
-      if (!fs) {
-        GetLogger()->error("Cannot create ShaderToy fragment shader: {}",
-                           vs.error().what());
-        IRIS_LOG_LEAVE();
-        return nullptr;
-      }
-
-      absl::FixedArray<iris::Renderer::Shader> shaders{
-        iris::Renderer::Shader{*vs, VK_SHADER_STAGE_VERTEX_BIT},
-        iris::Renderer::Shader{*fs, VK_SHADER_STAGE_FRAGMENT_BIT},
-      };
-
-      absl::FixedArray<VkPushConstantRange> pushConstantRanges(1);
-      pushConstantRanges[0] = {VK_SHADER_STAGE_VERTEX_BIT |
-                                 VK_SHADER_STAGE_FRAGMENT_BIT,
-                               0, sizeof(PushConstants)};
-
-      VkPipelineInputAssemblyStateCreateInfo inputAssemblyStateCI = {};
-      inputAssemblyStateCI.sType =
-        VK_STRUCTURE_TYPE_PIPELINE_INPUT_ASSEMBLY_STATE_CREATE_INFO;
-      inputAssemblyStateCI.topology = VK_PRIMITIVE_TOPOLOGY_TRIANGLE_LIST;
-
-      // The viewport and scissor are specified later as dynamic states
-      VkPipelineViewportStateCreateInfo viewportStateCI = {};
-      viewportStateCI.sType =
-        VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_STATE_CREATE_INFO;
-      viewportStateCI.viewportCount = 1;
-      viewportStateCI.scissorCount = 1;
-
-      VkPipelineRasterizationStateCreateInfo rasterizationStateCI = {};
-      rasterizationStateCI.sType =
-        VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_CREATE_INFO;
-      rasterizationStateCI.polygonMode = VK_POLYGON_MODE_FILL;
-      rasterizationStateCI.cullMode = VK_CULL_MODE_FRONT_BIT;
-      rasterizationStateCI.frontFace = VK_FRONT_FACE_COUNTER_CLOCKWISE;
-      rasterizationStateCI.lineWidth = 1.f;
-
-      VkPipelineMultisampleStateCreateInfo multisampleStateCI = {};
-      multisampleStateCI.sType =
-        VK_STRUCTURE_TYPE_PIPELINE_MULTISAMPLE_STATE_CREATE_INFO;
-      multisampleStateCI.rasterizationSamples = VK_SAMPLE_COUNT_4_BIT;
-      multisampleStateCI.minSampleShading = 1.f;
-
-      VkPipelineDepthStencilStateCreateInfo depthStencilStateCI = {};
-      depthStencilStateCI.sType =
-        VK_STRUCTURE_TYPE_PIPELINE_DEPTH_STENCIL_STATE_CREATE_INFO;
-
-      absl::FixedArray<VkPipelineColorBlendAttachmentState>
-        colorBlendAttachmentStates(1);
-      colorBlendAttachmentStates[0] = {
-        VK_FALSE,                            // blendEnable
-        VK_BLEND_FACTOR_SRC_ALPHA,           // srcColorBlendFactor
-        VK_BLEND_FACTOR_ONE_MINUS_SRC_ALPHA, // dstColorBlendFactor
-        VK_BLEND_OP_ADD,                     // colorBlendOp
-        VK_BLEND_FACTOR_ONE_MINUS_SRC_ALPHA, // srcAlphaBlendFactor
-        VK_BLEND_FACTOR_ZERO,                // dstAlphaBlendFactor
-        VK_BLEND_OP_ADD,                     // alphaBlendOp
-        VK_COLOR_COMPONENT_R_BIT | VK_COLOR_COMPONENT_G_BIT |
-          VK_COLOR_COMPONENT_B_BIT | VK_COLOR_COMPONENT_A_BIT // colorWriteMask
-      };
-
-      absl::FixedArray<VkDynamicState> dynamicStates{VK_DYNAMIC_STATE_VIEWPORT,
-                                                     VK_DYNAMIC_STATE_SCISSOR};
-
-      if (auto pl = iris::Renderer::CreateGraphicsPipeline(
-            {}, pushConstantRanges, shaders, {}, {}, inputAssemblyStateCI,
-            viewportStateCI, rasterizationStateCI, multisampleStateCI,
-            depthStencilStateCI, colorBlendAttachmentStates, dynamicStates, 0,
-            "iris-shadertoy::Renderable::Pipeline")) {
-        std::tie(renderable.pipelineLayout, renderable.pipeline) = *pl;
-      } else {
-        GetLogger()->error("Cannot create ShaderToy pipeline: {}",
-                           pl.error().what());
-        IRIS_LOG_LEAVE();
-        return nullptr;
-      }
-
-      renderable.numVertices = 3;
-      iris::Renderer::AddRenderable(renderable);
-
-      IRIS_LOG_LEAVE();
-      return nullptr;
-    } // execute
-
-  private:
-    std::vector<std::string> urls_;
-
-    // this throws
-    std::string GetCode(web::http::uri const& uri) {
-      std::string code;
-
-      web::http::client::http_client client(uri.to_string());
-      client.request(web::http::methods::GET)
-        .then([&](web::http::http_response response)
-                -> pplx::task<web::json::value> {
-          GetLogger()->debug(
-            "LoadShaderToy::LoadTask::GetCode: response status_code: {}",
-            response.status_code());
-          return response.extract_json();
-        })
-        .then([&](web::json::value json) {
-          auto&& renderpass =
-            json.at(_XPLATSTR("Shader")).at(_XPLATSTR("renderpass")).at(0);
-
-          if (renderpass.at(_XPLATSTR("inputs")).size() > 0) {
-            throw std::runtime_error("inputs are not yet implemented");
-          } else if (renderpass.at(_XPLATSTR("type")).as_string() !=
-                     _XPLATSTR("image")) {
-            throw std::runtime_error(
-              "non-image outputs are not yet implemented");
-          }
-
-#if PLATFORM_WINDOWS
-          code =
-            wstring_to_string(renderpass.at(_XPLATSTR("code")).as_string());
-#else
-          code = renderpass.at(_XPLATSTR("code")).as_string();
-#endif
-        })
-        .wait();
-
-      return code;
-    }; // GetCode
-  }; // class LoadTask
-
-  try {
-    LoadTask* task = new (tbb::task::allocate_root()) LoadTask(shaderToy);
-    tbb::task::enqueue(*task);
-  } catch (std::exception const& e) {
-    GetLogger()->warn("Loading shadertoy failed: {}", e.what());
+  switch (messageSeverity) {
+  case VK_DEBUG_UTILS_MESSAGE_SEVERITY_VERBOSE_BIT_EXT:
+    if (objNames.empty()) {
+      GetLogger()->trace(msg);
+    } else {
+      GetLogger()->trace("{} Objects: ({})", msg, objNames);
+    }
+    break;
+  case VK_DEBUG_UTILS_MESSAGE_SEVERITY_INFO_BIT_EXT:
+    if (objNames.empty()) {
+      GetLogger()->info(msg);
+    } else {
+      GetLogger()->info("{} Objects: ({})", msg, objNames);
+    }
+    break;
+  case VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT:
+    if (objNames.empty()) {
+      GetLogger()->warn(msg);
+    } else {
+      GetLogger()->warn("{} Objects: ({})", msg, objNames);
+    }
+    break;
+  case VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT:
+    if (objNames.empty()) {
+      GetLogger()->error(msg);
+    } else {
+      GetLogger()->error("{} Objects: ({})", msg, objNames);
+    }
+    break;
+  default:
+    GetLogger()->error("Unhandled VkDebugUtilsMessengerSeverityFlagBitsEXT: {}",
+                       messageSeverity);
+    if (objNames.empty()) {
+      GetLogger()->error(msg);
+    } else {
+      GetLogger()->error("{} Objects: ({})", msg, objNames);
+    }
+    break;
   }
 
-  IRIS_LOG_LEAVE();
-} // LoadShaderToy
-
-  static VKAPI_ATTR VkBool32 VKAPI_CALL
-  DebugUtilsMessengerCallback(
-    VkDebugUtilsMessageSeverityFlagBitsEXT messageSeverity,
-    VkDebugUtilsMessageTypeFlagsEXT messageTypes,
-    VkDebugUtilsMessengerCallbackDataEXT const* pCallbackData, void*) {
-    using namespace std::string_literals;
-
-    fmt::memory_buffer buf;
-    fmt::format_to(
-      buf, "{}: {}",
-      to_string(static_cast<VkDebugUtilsMessageTypeFlagBitsEXT>(messageTypes)),
-      pCallbackData->pMessage);
-    std::string const msg(buf.data(), buf.size());
-
-    buf.clear();
-    for (uint32_t i = 0; i < pCallbackData->objectCount; ++i) {
-      if (pCallbackData->pObjects[i].pObjectName) {
-        fmt::format_to(buf, "{}, ", pCallbackData->pObjects[i].pObjectName);
-      }
-    }
-    std::string const objNames(buf.data(),
-                               buf.size() == 0 ? 0 : buf.size() - 2);
-
-    switch (messageSeverity) {
-    case VK_DEBUG_UTILS_MESSAGE_SEVERITY_VERBOSE_BIT_EXT:
-      if (objNames.empty()) {
-        GetLogger()->trace(msg);
-      } else {
-        GetLogger()->trace("{} Objects: ({})", msg, objNames);
-      }
-      break;
-    case VK_DEBUG_UTILS_MESSAGE_SEVERITY_INFO_BIT_EXT:
-      if (objNames.empty()) {
-        GetLogger()->info(msg);
-      } else {
-        GetLogger()->info("{} Objects: ({})", msg, objNames);
-      }
-      break;
-    case VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT:
-      if (objNames.empty()) {
-        GetLogger()->warn(msg);
-      } else {
-        GetLogger()->warn("{} Objects: ({})", msg, objNames);
-      }
-      break;
-    case VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT:
-      if (objNames.empty()) {
-        GetLogger()->error(msg);
-      } else {
-        GetLogger()->error("{} Objects: ({})", msg, objNames);
-      }
-      break;
-    default:
-      GetLogger()->error(
-        "Unhandled VkDebugUtilsMessengerSeverityFlagBitsEXT: {}",
-        messageSeverity);
-      if (objNames.empty()) {
-        GetLogger()->error(msg);
-      } else {
-        GetLogger()->error("{} Objects: ({})", msg, objNames);
-      }
-      break;
-    }
-
-    GetLogger()->flush();
-    return VK_FALSE;
-  } // DebugUtilsMessengerCallback
+  GetLogger()->flush();
+  return VK_FALSE;
+} // DebugUtilsMessengerCallback
 
 } // namespace iris::Renderer
 
@@ -1685,7 +1329,7 @@ void iris::Renderer::EndFrame(
     std::vector<Component::Renderable> renderables = sRenderables();
     for (auto&& renderable : renderables) {
       VkCommandBuffer commandBuffer =
-        Render(renderable, &window.viewport, &window.scissor);
+        RenderRenderable(renderable, &window.viewport, &window.scissor);
       vkCmdExecuteCommands(frame.commandBuffer, 1, &commandBuffer);
     }
 
@@ -1825,7 +1469,9 @@ iris::Renderer::Control(iris::Control::Control const& controlMessage) noexcept {
     CreateEmplaceWindow(controlMessage.window());
     break;
   case iris::Control::Control_Type_SHADERTOY:
-    LoadShaderToy(controlMessage.shadertoy());
+    for (int i = 0; i < controlMessage.shadertoy().url_size(); ++i) {
+      sIOContinuations.push(io::LoadShaderToy(controlMessage.shadertoy().url(i)));
+    }
     break;
   default:
     GetLogger()->error("Unsupported controlMessage message type {}",
@@ -1841,459 +1487,3 @@ iris::Renderer::Control(iris::Control::Control const& controlMessage) noexcept {
   IRIS_LOG_LEAVE();
   return {};
 } // iris::Renderer::Control
-
-tl::expected<absl::FixedArray<VkCommandBuffer>, std::system_error>
-iris::Renderer::AllocateCommandBuffers(VkCommandBufferLevel level,
-                                       std::uint32_t count) noexcept {
-  IRIS_LOG_ENTER();
-  Expects(sDevice != VK_NULL_HANDLE);
-  Expects(count > 0);
-
-  VkCommandBufferAllocateInfo commandBufferAI = {};
-  commandBufferAI.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-  commandBufferAI.commandPool = sGraphicsCommandPools[0];
-  commandBufferAI.level = level;
-  commandBufferAI.commandBufferCount = count;
-
-  absl::FixedArray<VkCommandBuffer> commandBuffers(count);
-  if (auto result = vkAllocateCommandBuffers(sDevice, &commandBufferAI,
-                                             commandBuffers.data());
-      result != VK_SUCCESS) {
-    return tl::unexpected(std::system_error(make_error_code(result),
-                                            "Cannot allocate command buffers"));
-  }
-
-  IRIS_LOG_LEAVE();
-  return commandBuffers;
-} // iris::Renderer::AllocateCommandBuffers
-
-#if PLATFORM_COMPILER_GCC
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wshadow"
-#endif
-#include "SPIRV/GLSL.std.450.h"
-#include "SPIRV/GlslangToSpv.h"
-#include "glslang/Public/ShaderLang.h"
-#if PLATFORM_COMPILER_GCC
-#pragma GCC diagnostic pop
-#endif
-
-namespace iris::Renderer {
-
-const TBuiltInResource DefaultTBuiltInResource = {
-  /* .MaxLights = */ 32,
-  /* .MaxClipPlanes = */ 6,
-  /* .MaxTextureUnits = */ 32,
-  /* .MaxTextureCoords = */ 32,
-  /* .MaxVertexAttribs = */ 64,
-  /* .MaxVertexUniformComponents = */ 4096,
-  /* .MaxVaryingFloats = */ 64,
-  /* .MaxVertexTextureImageUnits = */ 32,
-  /* .MaxCombinedTextureImageUnits = */ 80,
-  /* .MaxTextureImageUnits = */ 32,
-  /* .MaxFragmentUniformComponents = */ 4096,
-  /* .MaxDrawBuffers = */ 32,
-  /* .MaxVertexUniformVectors = */ 128,
-  /* .MaxVaryingVectors = */ 8,
-  /* .MaxFragmentUniformVectors = */ 16,
-  /* .MaxVertexOutputVectors = */ 16,
-  /* .MaxFragmentInputVectors = */ 15,
-  /* .MinProgramTexelOffset = */ -8,
-  /* .MaxProgramTexelOffset = */ 7,
-  /* .MaxClipDistances = */ 8,
-  /* .MaxComputeWorkGroupCountX = */ 65535,
-  /* .MaxComputeWorkGroupCountY = */ 65535,
-  /* .MaxComputeWorkGroupCountZ = */ 65535,
-  /* .MaxComputeWorkGroupSizeX = */ 1024,
-  /* .MaxComputeWorkGroupSizeY = */ 1024,
-  /* .MaxComputeWorkGroupSizeZ = */ 64,
-  /* .MaxComputeUniformComponents = */ 1024,
-  /* .MaxComputeTextureImageUnits = */ 16,
-  /* .MaxComputeImageUniforms = */ 8,
-  /* .MaxComputeAtomicCounters = */ 8,
-  /* .MaxComputeAtomicCounterBuffers = */ 1,
-  /* .MaxVaryingComponents = */ 60,
-  /* .MaxVertexOutputComponents = */ 64,
-  /* .MaxGeometryInputComponents = */ 64,
-  /* .MaxGeometryOutputComponents = */ 128,
-  /* .MaxFragmentInputComponents = */ 128,
-  /* .MaxImageUnits = */ 8,
-  /* .MaxCombinedImageUnitsAndFragmentOutputs = */ 8,
-  /* .MaxCombinedShaderOutputResources = */ 8,
-  /* .MaxImageSamples = */ 0,
-  /* .MaxVertexImageUniforms = */ 0,
-  /* .MaxTessControlImageUniforms = */ 0,
-  /* .MaxTessEvaluationImageUniforms = */ 0,
-  /* .MaxGeometryImageUniforms = */ 0,
-  /* .MaxFragmentImageUniforms = */ 8,
-  /* .MaxCombinedImageUniforms = */ 8,
-  /* .MaxGeometryTextureImageUnits = */ 16,
-  /* .MaxGeometryOutputVertices = */ 256,
-  /* .MaxGeometryTotalOutputComponents = */ 1024,
-  /* .MaxGeometryUniformComponents = */ 1024,
-  /* .MaxGeometryVaryingComponents = */ 64,
-  /* .MaxTessControlInputComponents = */ 128,
-  /* .MaxTessControlOutputComponents = */ 128,
-  /* .MaxTessControlTextureImageUnits = */ 16,
-  /* .MaxTessControlUniformComponents = */ 1024,
-  /* .MaxTessControlTotalOutputComponents = */ 4096,
-  /* .MaxTessEvaluationInputComponents = */ 128,
-  /* .MaxTessEvaluationOutputComponents = */ 128,
-  /* .MaxTessEvaluationTextureImageUnits = */ 16,
-  /* .MaxTessEvaluationUniformComponents = */ 1024,
-  /* .MaxTessPatchComponents = */ 120,
-  /* .MaxPatchVertices = */ 32,
-  /* .MaxTessGenLevel = */ 64,
-  /* .MaxViewports = */ 16,
-  /* .MaxVertexAtomicCounters = */ 0,
-  /* .MaxTessControlAtomicCounters = */ 0,
-  /* .MaxTessEvaluationAtomicCounters = */ 0,
-  /* .MaxGeometryAtomicCounters = */ 0,
-  /* .MaxFragmentAtomicCounters = */ 8,
-  /* .MaxCombinedAtomicCounters = */ 8,
-  /* .MaxAtomicCounterBindings = */ 1,
-  /* .MaxVertexAtomicCounterBuffers = */ 0,
-  /* .MaxTessControlAtomicCounterBuffers = */ 0,
-  /* .MaxTessEvaluationAtomicCounterBuffers = */ 0,
-  /* .MaxGeometryAtomicCounterBuffers = */ 0,
-  /* .MaxFragmentAtomicCounterBuffers = */ 1,
-  /* .MaxCombinedAtomicCounterBuffers = */ 1,
-  /* .MaxAtomicCounterBufferSize = */ 16384,
-  /* .MaxTransformFeedbackBuffers = */ 4,
-  /* .MaxTransformFeedbackInterleavedComponents = */ 64,
-  /* .MaxCullDistances = */ 8,
-  /* .MaxCombinedClipAndCullDistances = */ 8,
-  /* .MaxSamples = */ 4,
-  /* .maxMeshOutputVerticesNV = */ 256,
-  /* .maxMeshOutputPrimitivesNV = */ 512,
-  /* .maxMeshWorkGroupSizeX_NV = */ 32,
-  /* .maxMeshWorkGroupSizeY_NV = */ 1,
-  /* .maxMeshWorkGroupSizeZ_NV = */ 1,
-  /* .maxTaskWorkGroupSizeX_NV = */ 32,
-  /* .maxTaskWorkGroupSizeY_NV = */ 1,
-  /* .maxTaskWorkGroupSizeZ_NV = */ 1,
-  /* .maxMeshViewCountNV = */ 4,
-
-  /* .limits = */
-  {
-    /* .nonInductiveForLoops = */ true,
-    /* .whileLoops = */ true,
-    /* .doWhileLoops = */ true,
-    /* .generalUniformIndexing = */ true,
-    /* .generalAttributeMatrixVectorIndexing = */ true,
-    /* .generalVaryingIndexing = */ true,
-    /* .generalSamplerIndexing = */ true,
-    /* .generalVariableIndexing = */ true,
-    /* .generalConstantMatrixVectorIndexing = */ true,
-  }};
-
-class DirStackIncluder : public glslang::TShader::Includer {
-public:
-  DirStackIncluder() noexcept = default;
-
-  virtual IncludeResult* includeLocal(char const* headerName,
-                                      char const* includerName,
-                                      std::size_t inclusionDepth) override {
-    return readLocalPath(headerName, includerName, inclusionDepth);
-  }
-
-  virtual IncludeResult* includeSystem(char const* headerName,
-                                       char const* includerName
-                                       [[maybe_unused]],
-                                       std::size_t inclusionDepth
-                                       [[maybe_unused]]) override {
-    return readSystemPath(headerName);
-  }
-
-  virtual void releaseInclude(IncludeResult* result) override {
-    if (result) {
-      delete[] static_cast<char*>(result->userData);
-      delete result;
-    }
-  }
-
-  virtual void pushExternalLocalDirectory(std::string const& dir) {
-    dirStack_.push_back(dir);
-    numExternalLocalDirs_ = dirStack_.size();
-  }
-
-private:
-  std::vector<std::string> dirStack_{};
-  int numExternalLocalDirs_{0};
-
-  virtual IncludeResult* readLocalPath(std::string const& headerName,
-                                       std::string const& includerName,
-                                       int depth) {
-    // Discard popped include directories, and
-    // initialize when at parse-time first level.
-    dirStack_.resize(depth + numExternalLocalDirs_);
-
-    if (depth == 1) dirStack_.back() = getDirectory(includerName);
-
-    // Find a directory that works, using a reverse search of the include stack.
-    for (auto& dir : dirStack_) {
-      std::string path = dir + "/"s + headerName;
-      std::replace(path.begin(), path.end(), '\\', '/');
-      std::ifstream ifs(path.c_str(),
-                        std::ios_base::binary | std::ios_base::ate);
-      if (ifs) {
-        dirStack_.push_back(getDirectory(path));
-        return newIncludeResult(path, ifs, ifs.tellg());
-      }
-    }
-
-    return nullptr;
-  }
-
-  virtual IncludeResult* readSystemPath(char const*) const {
-    GetLogger()->error("including system headers not implemented");
-    return nullptr;
-  }
-
-  virtual IncludeResult* newIncludeResult(std::string const& path,
-                                          std::ifstream& ifs,
-                                          int length) const {
-    char* content = new char[length];
-    ifs.seekg(0, ifs.beg);
-    ifs.read(content, length);
-    return new IncludeResult(path.c_str(), content, length, content);
-  }
-
-  // If no path markers, return current working directory.
-  // Otherwise, strip file name and return path leading up to it.
-  virtual std::string getDirectory(const std::string path) const {
-    size_t last = path.find_last_of("/\\");
-    return last == std::string::npos ? "." : path.substr(0, last);
-  }
-}; // class DirStackIncluder
-
-[[nodiscard]] static tl::expected<std::vector<std::uint32_t>, std::string>
-CompileShader(std::string_view source, VkShaderStageFlagBits shaderStage,
-              filesystem::path const& path,
-              gsl::span<std::string> macroDefinitions [[maybe_unused]],
-              std::string const& entryPoint) {
-  IRIS_LOG_ENTER();
-  Expects(source.size() > 0);
-
-  auto const lang = [&shaderStage]() {
-    if ((shaderStage & VK_SHADER_STAGE_VERTEX_BIT)) {
-      return EShLanguage::EShLangVertex;
-    } else if ((shaderStage & VK_SHADER_STAGE_FRAGMENT_BIT)) {
-      return EShLanguage::EShLangFragment;
-    } else {
-      GetLogger()->critical("Unhandled shaderStage: {}", shaderStage);
-      std::terminate();
-    }
-  }();
-
-  char const* strings[] = {source.data()};
-  int lengths[] = {static_cast<int>(source.size())};
-  char const* names[] = {path.string().c_str()};
-
-  glslang::TShader shader(lang);
-  shader.setStringsWithLengthsAndNames(strings, lengths, names, 1);
-  shader.setEntryPoint(entryPoint.c_str());
-  shader.setEnvInput(glslang::EShSource::EShSourceGlsl, lang,
-                     glslang::EShClient::EShClientVulkan, 101);
-  shader.setEnvClient(glslang::EShClient::EShClientVulkan,
-                      glslang::EShTargetClientVersion::EShTargetVulkan_1_1);
-  shader.setEnvTarget(glslang::EShTargetLanguage::EShTargetSpv,
-                      glslang::EShTargetLanguageVersion::EShTargetSpv_1_0);
-
-  DirStackIncluder includer;
-  includer.pushExternalLocalDirectory(kIRISContentDirectory);
-
-  if (!shader.parse(&DefaultTBuiltInResource, 1, false,
-                    EShMessages::EShMsgDefault, includer)) {
-    return tl::unexpected(std::string(shader.getInfoLog()));
-  }
-
-  glslang::TProgram program;
-  program.addShader(&shader);
-
-  if (!program.link(EShMessages::EShMsgDefault)) {
-    return tl::unexpected(std::string(program.getInfoLog()));
-  }
-
-  if (auto glsl = program.getIntermediate(lang)) {
-    glslang::SpvOptions options;
-    options.validate = true;
-#ifndef NDEBUG
-    options.generateDebugInfo = true;
-#endif
-
-    spv::SpvBuildLogger logger;
-    std::vector<std::uint32_t> code;
-    glslang::GlslangToSpv(*glsl, code, &logger, &options);
-
-    Ensures(code.size() > 0);
-    IRIS_LOG_LEAVE();
-    return code;
-  } else {
-    return tl::unexpected(std::string(
-      "cannot get glsl intermediate representation of compiled shader"));
-  }
-} // CompileShader
-
-} // namespace iris::Renderer
-
-tl::expected<VkShaderModule, std::system_error>
-iris::Renderer::CompileShaderFromSource(std::string_view source,
-                                        VkShaderStageFlagBits stage,
-                                        std::string name) noexcept {
-  IRIS_LOG_ENTER();
-  Expects(sDevice != VK_NULL_HANDLE);
-  Expects(source.size() > 0);
-
-  VkShaderModule module{VK_NULL_HANDLE};
-
-  auto code = CompileShader(source, stage, "<inline>", {}, "main");
-  if (!code) {
-    IRIS_LOG_LEAVE();
-    return tl::unexpected(
-      std::system_error(Error::kShaderCompileFailed, code.error()));
-  }
-
-  VkShaderModuleCreateInfo shaderModuleCI = {};
-  shaderModuleCI.sType = VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO;
-  // codeSize is count of bytes, not count of words (which is what size() is)
-  shaderModuleCI.codeSize = gsl::narrow_cast<std::uint32_t>(code->size()) * 4u;
-  shaderModuleCI.pCode = code->data();
-
-  if (auto result =
-        vkCreateShaderModule(sDevice, &shaderModuleCI, nullptr, &module);
-      result != VK_SUCCESS) {
-    IRIS_LOG_LEAVE();
-    return tl::unexpected(std::system_error(make_error_code(result),
-                                            "Cannot create shader module"));
-  }
-
-  if (!name.empty()) {
-    NameObject(sDevice, VK_OBJECT_TYPE_SHADER_MODULE, module, name.c_str());
-  }
-
-  Ensures(module != VK_NULL_HANDLE);
-  IRIS_LOG_LEAVE();
-  return module;
-} // iris::Renderer::CompileShaderFromSource
-
-tl::expected<std::pair<VkPipelineLayout, VkPipeline>, std::system_error>
-iris::Renderer::CreateGraphicsPipeline(
-  gsl::span<const VkDescriptorSetLayout> descriptorSetLayouts,
-  gsl::span<const VkPushConstantRange> pushConstantRanges,
-  gsl::span<const Shader> shaders,
-  gsl::span<const VkVertexInputBindingDescription>
-    vertexInputBindingDescriptions,
-  gsl::span<const VkVertexInputAttributeDescription>
-    vertexInputAttributeDescriptions,
-  VkPipelineInputAssemblyStateCreateInfo inputAssemblyStateCI,
-  VkPipelineViewportStateCreateInfo viewportStateCI,
-  VkPipelineRasterizationStateCreateInfo rasterizationStateCI,
-  VkPipelineMultisampleStateCreateInfo multisampleStateCI,
-  VkPipelineDepthStencilStateCreateInfo depthStencilStateCI,
-  gsl::span<const VkPipelineColorBlendAttachmentState>
-    colorBlendAttachmentStates,
-  gsl::span<const VkDynamicState> dynamicStates,
-  std::uint32_t renderPassSubpass, std::string name) noexcept {
-  IRIS_LOG_ENTER();
-  Expects(sDevice != VK_NULL_HANDLE);
-  Expects(sRenderPass != VK_NULL_HANDLE);
-
-  VkPipelineLayout layout{VK_NULL_HANDLE};
-  VkPipeline pipeline{VK_NULL_HANDLE};
-
-  VkPipelineLayoutCreateInfo pipelineLayoutCI = {};
-  pipelineLayoutCI.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-  pipelineLayoutCI.setLayoutCount =
-    gsl::narrow_cast<std::uint32_t>(descriptorSetLayouts.size());
-  pipelineLayoutCI.pSetLayouts = descriptorSetLayouts.data();
-  pipelineLayoutCI.pushConstantRangeCount =
-    gsl::narrow_cast<std::uint32_t>(pushConstantRanges.size());
-  pipelineLayoutCI.pPushConstantRanges = pushConstantRanges.data();
-
-  if (auto result =
-        vkCreatePipelineLayout(sDevice, &pipelineLayoutCI, nullptr, &layout);
-      result != VK_SUCCESS) {
-    IRIS_LOG_LEAVE();
-    return tl::unexpected(std::system_error(make_error_code(result),
-                                            "Cannot create pipeline layout"));
-  }
-
-  absl::FixedArray<VkPipelineShaderStageCreateInfo> shaderStageCIs(
-    shaders.size());
-  std::transform(shaders.begin(), shaders.end(), shaderStageCIs.begin(),
-                 [](Shader const& shader) {
-                   return VkPipelineShaderStageCreateInfo{
-                     VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO,
-                     nullptr,
-                     0,
-                     shader.stage,
-                     shader.handle,
-                     "main",
-                     nullptr};
-                 });
-
-  VkPipelineVertexInputStateCreateInfo vertexInputStateCI = {};
-  vertexInputStateCI.sType =
-    VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_STATE_CREATE_INFO;
-  vertexInputStateCI.vertexBindingDescriptionCount =
-    gsl::narrow_cast<std::uint32_t>(vertexInputBindingDescriptions.size());
-  vertexInputStateCI.pVertexBindingDescriptions =
-    vertexInputBindingDescriptions.data();
-  vertexInputStateCI.vertexAttributeDescriptionCount =
-    gsl::narrow_cast<std::uint32_t>(vertexInputAttributeDescriptions.size());
-  vertexInputStateCI.pVertexAttributeDescriptions =
-    vertexInputAttributeDescriptions.data();
-
-  VkPipelineColorBlendStateCreateInfo colorBlendStateCI = {};
-  colorBlendStateCI.sType =
-    VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_STATE_CREATE_INFO;
-  colorBlendStateCI.attachmentCount =
-    gsl::narrow_cast<std::uint32_t>(colorBlendAttachmentStates.size());
-  colorBlendStateCI.pAttachments = colorBlendAttachmentStates.data();
-
-  VkPipelineDynamicStateCreateInfo dynamicStateCI = {};
-  dynamicStateCI.sType = VK_STRUCTURE_TYPE_PIPELINE_DYNAMIC_STATE_CREATE_INFO;
-  dynamicStateCI.dynamicStateCount =
-    gsl::narrow_cast<uint32_t>(dynamicStates.size());
-  dynamicStateCI.pDynamicStates = dynamicStates.data();
-
-  VkGraphicsPipelineCreateInfo graphicsPipelineCI = {};
-  graphicsPipelineCI.sType = VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO;
-  graphicsPipelineCI.stageCount = static_cast<uint32_t>(shaderStageCIs.size());
-  graphicsPipelineCI.pStages = shaderStageCIs.data();
-  graphicsPipelineCI.pVertexInputState = &vertexInputStateCI;
-  graphicsPipelineCI.pInputAssemblyState = &inputAssemblyStateCI;
-  graphicsPipelineCI.pViewportState = &viewportStateCI;
-  graphicsPipelineCI.pRasterizationState = &rasterizationStateCI;
-  graphicsPipelineCI.pMultisampleState = &multisampleStateCI;
-  graphicsPipelineCI.pDepthStencilState = &depthStencilStateCI;
-  graphicsPipelineCI.pColorBlendState = &colorBlendStateCI;
-  graphicsPipelineCI.pDynamicState = &dynamicStateCI;
-  graphicsPipelineCI.layout = layout;
-  graphicsPipelineCI.renderPass = sRenderPass;
-  graphicsPipelineCI.subpass = renderPassSubpass;
-
-  if (auto result = vkCreateGraphicsPipelines(
-        sDevice, VK_NULL_HANDLE, 1, &graphicsPipelineCI, nullptr, &pipeline);
-      result != VK_SUCCESS) {
-    IRIS_LOG_LEAVE();
-    return tl::unexpected(std::system_error(make_error_code(result),
-                                            "Cannot create graphics pipeline"));
-  }
-
-  if (!name.empty()) {
-    NameObject(sDevice, VK_OBJECT_TYPE_PIPELINE_LAYOUT, layout,
-               (name + ".layout").c_str());
-    NameObject(sDevice, VK_OBJECT_TYPE_PIPELINE, pipeline, name.c_str());
-  }
-
-  Ensures(layout != VK_NULL_HANDLE);
-  Ensures(pipeline != VK_NULL_HANDLE);
-  IRIS_LOG_LEAVE();
-  return std::make_pair(layout, pipeline);
-} // iris::Renderer::CreateGraphicsPipeline
-
-void iris::Renderer::AddRenderable(Component::Renderable renderable) noexcept {
-  sRenderables.push_back(std::move(renderable));
-} // AddRenderable
